+++ src/api/grid/iters.rs
@@ -9,14 +9,14 @@ impl<A: Arch> Grid<2, A> {
     pub fn x_iter(&self) -> RowIter<A> {
         let pos = self.config.position;
         let dim = self.config.grid_size;
-        RowIter::new(dim[0], dim[1], pos[0] as f32)
+        RowIter::new(dim[0], dim[1], pos[0])
     }
 
     #[inline(always)]
     pub fn y_iter(&self) -> SliceIter<A> {
         let pos = self.config.position;
         let dim = self.config.grid_size;
-        SliceIter::new(dim[0], dim[1], 1, pos[1] as f32)
+        SliceIter::new(dim[0], dim[1], 1, pos[1])
     }
 }
 
@@ -25,21 +25,21 @@ impl<A: Arch> Grid<3, A> {
     pub fn x_iter(&self) -> RowIter<A> {
         let pos = self.config.position;
         let dim = self.config.grid_size;
-        RowIter::new(dim[0], dim[1] * dim[2], pos[0] as f32)
+        RowIter::new(dim[0], dim[1] * dim[2], pos[0])
     }
 
     #[inline(always)]
     pub fn y_iter(&self) -> SliceIter<A> {
         let pos = self.config.position;
         let dim = self.config.grid_size;
-        SliceIter::new(dim[0], dim[1], dim[2], pos[1] as f32)
+        SliceIter::new(dim[0], dim[1], dim[2], pos[1])
     }
 
     #[inline(always)]
     pub fn z_iter(&self) -> SliceIter<A> {
         let pos = self.config.position;
         let dim = self.config.grid_size;
-        SliceIter::new(dim[0] * dim[1], dim[2], 1, pos[2] as f32)
+        SliceIter::new(dim[0] * dim[1], dim[2], 1, pos[2])
     }
 }
